[maze-generator] Biased generator
This generator is based on the "from solution" generator.
One of the issue of the "from solution" generator is to generate too long
solution, making it impossible to create relevant fake paths.
The first idea was to biased the generation of the solution, so that it
try to reach the end quickly.

Then to avoid having too simple mazes, fake paths close to the start point
will try to reach the end quickly while the ones close to the end will try
to reach the start.

The idea behing those choices is that human brain will certainly try to do
some kind of A* algorithm starting from both start and end.